[refactor] DB refactor to separate threads -> threadInfo + threadMessages
